fewshot-doc-classifier/requirements.txt
openai
langchain
tiktoken
python-dotenv

